+++ benchmarks/Kernels/Distance/SafeEuclideanBench.php
@@ -4,6 +4,7 @@ namespace Rubix\ML\Benchmarks\Kernels\Distance;
 
 use Rubix\ML\Datasets\Generators\Blob;
 use Rubix\ML\Kernels\Distance\SafeEuclidean;
+use Rubix\ML\Transformers\LambdaFunction;
 
 /**
  * @Groups({"DistanceKernels"})
@@ -32,8 +33,18 @@ class SafeEuclideanBench
     {
         $generator = new Blob([0, 0, 0, 0, 0, 0, 0, 0], 5.0);
 
-        $this->aSamples = $generator->generate(self::NUM_SAMPLES)->samples();
-        $this->bSamples = $generator->generate(self::NUM_SAMPLES)->samples();
+        $dropValues = new LambdaFunction((function ($sample) {
+            $sample[4] = rand(0, 5) === 0 ? NAN : $sample[4];
+            $sample[5] = rand(0, 10) === 0 ? NAN : $sample[5];
+        }));
+
+        $this->aSamples = $generator->generate(self::NUM_SAMPLES)
+            ->apply($dropValues)
+            ->samples();
+
+        $this->bSamples = $generator->generate(self::NUM_SAMPLES)
+            ->apply($dropValues)
+            ->samples();
 
         $this->kernel = new SafeEuclidean();
     }