+++ src/Datasets/Dataset.php
@@ -17,7 +17,6 @@ use ArrayAccess;
 use Countable;
 
 use function Rubix\ML\iterator_first;
-use function Rubix\ML\iterator_map;
 use function Rubix\ML\iterator_filter;
 use function Rubix\ML\array_transpose;
 use function count;
@@ -127,6 +126,18 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
     public function size() : int
     {
         return $this->numSamples() * $this->numFeatures();
+    }
+
+    /**
+     * Return the high-level data types of each column in the data table.
+     *
+     * @return list<\Rubix\ML\DataType>
+     */
+    public function types() : array
+    {
+        $firstRow = iterator_first($this);
+
+        return array_map([DataType::class, 'detect'], $firstRow);
     }
 
     /**
@@ -302,7 +313,7 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
     {
         if ($transformer instanceof Stateful) {
             if (!$transformer->fitted()) {
-                throw new RuntimeException('Transformer has not been fitted.');
+                throw new RuntimeException('Stateful transformer has not been fitted.');
             }
         }
 
@@ -338,11 +349,9 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
 
         $columns = array_transpose(iterator_to_array($this));
 
-        $types = iterator_map(iterator_first($this), [DataType::class, 'detect']);
-
         $stats = [];
 
-        foreach ($types as $offset => $type) {
+        foreach ($this->types() as $offset => $type) {
             $description = [
                 'offset' => $offset,
                 'type' => (string) $type,
@@ -354,20 +363,22 @@ abstract class Dataset implements ArrayAccess, IteratorAggregate, Countable
                 case DataType::CONTINUOUS:
                     [$mean, $variance] = Stats::meanVar($values);
 
-                    $quantiles = Stats::quantiles($values, [
+                    [$min, $p25, $median, $p75, $max] = Stats::quantiles($values, [
                         0.0, 0.25, 0.5, 0.75, 1.0,
                     ]);
 
                     $description += [
                         'mean' => $mean,
+                        'variance' => $variance,
                         'standard deviation' => sqrt($variance),
                         'skewness' => Stats::skewness($values, $mean),
                         'kurtosis' => Stats::kurtosis($values, $mean),
-                        'min' => $quantiles[0],
-                        '25%' => $quantiles[1],
-                        'median' => $quantiles[2],
-                        '75%' => $quantiles[3],
-                        'max' => $quantiles[4],
+                        'min' => $min,
+                        '25%' => $p25,
+                        'median' => $median,
+                        '75%' => $p75,
+                        'max' => $max,
+                        'range' => $max - $min,
                     ];
 
                     break;