+++ tests/Classifiers/LogitBoostTest.php
@@ -31,14 +31,14 @@ class LogitBoostTest extends TestCase
      *
      * @var int
      */
-    protected const TRAIN_SIZE = 350;
+    protected const TRAIN_SIZE = 512;
 
     /**
      * The number of samples in the validation set.
      *
      * @var int
      */
-    protected const TEST_SIZE = 20;
+    protected const TEST_SIZE = 256;
 
     /**
      * The minimum validation score required to pass the test.