+++ tests/Classifiers/OneVsRestTest.php
@@ -1,0 +1,169 @@
+<?php
+
+namespace Rubix\ML\Tests\Classifiers;
+
+use Rubix\ML\Learner;
+use Rubix\ML\Parallel;
+use Rubix\ML\DataType;
+use Rubix\ML\Estimator;
+use Rubix\ML\Persistable;
+use Rubix\ML\Probabilistic;
+use Rubix\ML\EstimatorType;
+use Rubix\ML\Backends\Serial;
+use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Classifiers\OneVsRest;
+use Rubix\ML\Classifiers\GaussianNB;
+use Rubix\ML\Datasets\Generators\Blob;
+use Rubix\ML\CrossValidation\Metrics\FBeta;
+use Rubix\ML\Datasets\Generators\Agglomerate;
+use Rubix\ML\Exceptions\RuntimeException;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @group Classifiers
+ * @covers \Rubix\ML\Classifiers\OneVsRest
+ */
+class OneVsRestTest extends TestCase
+{
+    /**
+     * The number of samples in the training set.
+     *
+     * @var int
+     */
+    protected const TRAIN_SIZE = 512;
+
+    /**
+     * The number of samples in the validation set.
+     *
+     * @var int
+     */
+    protected const TEST_SIZE = 256;
+
+    /**
+     * The minimum validation score required to pass the test.
+     *
+     * @var float
+     */
+    protected const MIN_SCORE = 0.9;
+
+    /**
+     * Constant used to see the random number generator.
+     *
+     * @var int
+     */
+    protected const RANDOM_SEED = 0;
+
+    /**
+     * @var \Rubix\ML\Datasets\Generators\Agglomerate
+     */
+    protected $generator;
+
+    /**
+     * @var \Rubix\ML\Classifiers\OneVsRest
+     */
+    protected $estimator;
+
+    /**
+     * @var \Rubix\ML\CrossValidation\Metrics\FBeta
+     */
+    protected $metric;
+
+    /**
+     * @before
+     */
+    protected function setUp() : void
+    {
+        $this->generator = new Agglomerate([
+            'red' => new Blob([255, 32, 0], 50.0),
+            'green' => new Blob([0, 128, 0], 10.0),
+            'blue' => new Blob([0, 32, 255], 30.0),
+        ], [0.5, 0.2, 0.3]);
+
+        $this->estimator = new OneVsRest(new GaussianNB());
+
+        $this->estimator->setBackend(new Serial());
+
+        $this->metric = new FBeta();
+
+        srand(self::RANDOM_SEED);
+    }
+
+    protected function assertPreConditions() : void
+    {
+        $this->assertFalse($this->estimator->trained());
+    }
+
+    /**
+     * @test
+     */
+    public function build() : void
+    {
+        $this->assertInstanceOf(OneVsRest::class, $this->estimator);
+        $this->assertInstanceOf(Estimator::class, $this->estimator);
+        $this->assertInstanceOf(Learner::class, $this->estimator);
+        $this->assertInstanceOf(Probabilistic::class, $this->estimator);
+        $this->assertInstanceOf(Parallel::class, $this->estimator);
+        $this->assertInstanceOf(Persistable::class, $this->estimator);
+    }
+
+    /**
+     * @test
+     */
+    public function type() : void
+    {
+        $this->assertEquals(EstimatorType::classifier(), $this->estimator->type());
+    }
+
+    /**
+     * @test
+     */
+    public function compatibility() : void
+    {
+        $expected = [
+            DataType::continuous(),
+        ];
+
+        $this->assertEquals($expected, $this->estimator->compatibility());
+    }
+
+    /**
+     * @test
+     */
+    public function params() : void
+    {
+        $expected = [
+            'base' => new GaussianNB(),
+        ];
+
+        $this->assertEquals($expected, $this->estimator->params());
+    }
+
+    /**
+     * @test
+     */
+    public function trainPredictProba() : void
+    {
+        $training = $this->generator->generate(self::TRAIN_SIZE);
+        $testing = $this->generator->generate(self::TEST_SIZE);
+
+        $this->estimator->train($training);
+
+        $this->assertTrue($this->estimator->trained());
+
+        $predictions = $this->estimator->predict($testing);
+
+        $score = $this->metric->score($predictions, $testing->labels());
+
+        $this->assertGreaterThanOrEqual(self::MIN_SCORE, $score);
+    }
+
+    /**
+     * @test
+     */
+    public function predictUntrained() : void
+    {
+        $this->expectException(RuntimeException::class);
+
+        $this->estimator->predict(Unlabeled::quick());
+    }
+}