+++ tests/NeuralNet/Layers/BinaryTest.php
@@ -83,7 +83,7 @@ class BinaryTest extends TestCase
         $forward = $this->layer->forward($this->input);
 
         $this->assertInstanceOf(Matrix::class, $forward);
-        $this->assertEquals($expected, $forward->asArray());
+        $this->assertEqualsWithDelta($expected, $forward->asArray(), 1e-8);
 
         [$computation, $loss] = $this->layer->back($this->labels, $this->optimizer);
 
@@ -97,7 +97,7 @@ class BinaryTest extends TestCase
         ];
 
         $this->assertInstanceOf(Matrix::class, $gradient);
-        $this->assertEquals($expected, $gradient->asArray());
+        $this->assertEqualsWithDelta($expected, $gradient->asArray(), 1e-8);
 
         $expected = [
             [0.7310585786300049, 0.9241418199787566, 0.47502081252106],
@@ -106,6 +106,6 @@ class BinaryTest extends TestCase
         $infer = $this->layer->infer($this->input);
 
         $this->assertInstanceOf(Matrix::class, $infer);
-        $this->assertEquals($expected, $infer->asArray());
+        $this->assertEqualsWithDelta($expected, $infer->asArray(), 1e-8);
     }
 }