+++ tests/Regressors/AdalineTest.php
@@ -40,7 +40,7 @@ class AdalineTest extends TestCase
      *
      * @var int
      */
-    protected const TEST_SIZE = 128;
+    protected const TEST_SIZE = 256;
 
     /**
      * The minimum validation score required to pass the test.