+++ tests/Regressors/KNNRegressorTest.php
@@ -36,7 +36,7 @@ class KNNRegressorTest extends TestCase
      *
      * @var int
      */
-    protected const TEST_SIZE = 128;
+    protected const TEST_SIZE = 256;
 
     /**
      * The minimum validation score required to pass the test.
@@ -74,7 +74,7 @@ class KNNRegressorTest extends TestCase
     {
         $this->generator = new HalfMoon(4.0, -7.0, 1.0, 90, 0.25);
 
-        $this->estimator = new KNNRegressor(5, true, new Minkowski(3.0));
+        $this->estimator = new KNNRegressor(10, true, new Minkowski(3.0));
 
         $this->metric = new RSquared();
 
@@ -134,7 +134,7 @@ class KNNRegressorTest extends TestCase
     public function params() : void
     {
         $expected = [
-            'k' => 5,
+            'k' => 10,
             'weighted' => true,
             'kernel' => new Minkowski(3.0),
         ];