+++ tests/Regressors/MLPRegressorTest.php
@@ -44,7 +44,7 @@ class MLPRegressorTest extends TestCase
      *
      * @var int
      */
-    protected const TEST_SIZE = 128;
+    protected const TEST_SIZE = 256;
 
     /**
      * The minimum validation score required to pass the test.